requirements.txt
asgiref             3.8.1
Django              5.1.3
djangorestframework 3.15.2
pip                 21.2.4
setuptools          58.1.0
sqlparse            0.5.2
typing_extensions   4.12.2
tzdata              2024.2